fix date (long, not int), 1.11

+++ src/com/example/smsinformer/CentreonMessages.java
@@ -44,7 +44,7 @@ public class CentreonMessages {
 
 	}
 
-	CentreonMessage getNew(int time) {
+	CentreonMessage getNew(long time) {
 		if (counter != 0 && counter + 1 < messages.size()) {
 			counter++;
 			return messages.get(counter);
